Refactor imports in render-widget component and update TypeScript paths
