+++ packages/vscode-webui/tsconfig.json
@@ -27,7 +27,6 @@
       "@/lib/*": ["./src/lib/*"],
       "@/features/approval": ["./src/features/approval"],
       "@/features/chat": ["./src/features/chat"],
-      "@/features/chat/*": ["./src/features/chat/*"],
       "@/features/settings": ["./src/features/settings"],
       "@/features/share": ["./src/features/share"],
       "@/features/todo": ["./src/features/todo"],
